+++ src/main/java/project/seatsence/src/store/api/AdminStoreApi.java
@@ -1,5 +1,6 @@
 package project.seatsence.src.store.api;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.HashMap;
@@ -12,11 +13,12 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import project.seatsence.src.store.domain.Store;
+import project.seatsence.src.store.domain.StoreMember;
 import project.seatsence.src.store.dto.AdminStoreMapper;
-import project.seatsence.src.store.dto.request.AdminStoreCreateRequest;
-import project.seatsence.src.store.dto.request.AdminStoreFormCreateRequest;
-import project.seatsence.src.store.dto.request.AdminStoreUpdateRequest;
+import project.seatsence.src.store.dto.StoreMemberMapper;
+import project.seatsence.src.store.dto.request.*;
 import project.seatsence.src.store.dto.response.*;
+import project.seatsence.src.store.service.StoreMemberService;
 import project.seatsence.src.store.service.StoreService;
 import project.seatsence.src.store.service.StoreSpaceService;
 
@@ -31,6 +33,7 @@ public class AdminStoreApi {
     private final StoreService storeService;
     private final StoreSpaceService storeSpaceService;
     private final AdminStoreMapper adminStoreMapper;
+    private final StoreMemberService storeMemberService;
 
     @Operation(summary = "admin이 소유한 모든 가게 정보 가져오기")
     @GetMapping("/owned/{user-id}")
@@ -86,5 +89,48 @@ public class AdminStoreApi {
             @PathVariable("store-id") Long storeId,
             @RequestBody List<@Valid AdminStoreFormCreateRequest> adminStoreFormCreateRequestList) {
         storeSpaceService.save(storeId, adminStoreFormCreateRequestList);
+    }
+
+    @Operation(summary = "admin 직원 등록")
+    @PostMapping("/member-registration/{store-id}")
+    public void registerStoreMember(
+            @PathVariable("store-id") Long storeId,
+            @Valid @RequestBody StoreMemberRegistrationRequest storeMemberRegistrationRequest)
+            throws JsonProcessingException {
+
+        storeMemberService.storeMemberRegistration(storeId, storeMemberRegistrationRequest);
+    }
+
+    @Operation(summary = "가게 직원 리스트")
+    @GetMapping("/member-registration/{store-id}")
+    public StoreMemberListResponse getStoreMember(@PathVariable("store-id") Long storeId) {
+
+        List<StoreMember> storeMembers = storeMemberService.findAllByStoreIdAndPosition(storeId);
+
+        List<StoreMemberListResponse.StoreMemberResponse> storeMemberResponseList =
+                storeMembers.stream()
+                        .map(StoreMemberMapper::toStoreMemberResponse)
+                        .collect(Collectors.toList());
+
+        return StoreMemberListResponse.builder()
+                .storeMemberResponseList(storeMemberResponseList)
+                .build();
+    }
+
+    @Operation(summary = "직원 권한 수정")
+    @PatchMapping("/member-registration/{store-id}")
+    public void updateStoreMember(
+            @PathVariable("store-id") Long storeId,
+            @Valid @RequestBody StoreMemberUpdateRequest storeMemberUpdateRequest)
+            throws JsonProcessingException {
+        storeMemberService.update(storeId, storeMemberUpdateRequest);
+    }
+
+    @Operation(summary = "직원 삭제")
+    @DeleteMapping("/member-registration/{store-id}")
+    public void deleteStoreMember(
+            @PathVariable("store-id") Long storeId,
+            @Valid @RequestParam("member-id") Long storeMemberAuthorityId) {
+        storeMemberService.delete(storeMemberAuthorityId);
     }
 }
